コンポーネントに持たせすぎた役割を分離
