+++ src/pages/edit.tsx
@@ -1,11 +1,13 @@
+import { useRouter } from "next/router";
 import React from "react";
-import { useSelector } from "react-redux";
+import { useDispatch, useSelector } from "react-redux";
 import styled from "styled-components";
 
 import EditForm from "../components/EditForm";
 import Heading from "../components/Heading";
 import Layout from "../components/Layout";
 import { RootState } from "../store";
+import { createDraft } from "../store/diary/actions";
 
 const StyledLayout = styled(Layout)`
   > div {
@@ -20,12 +22,26 @@ const StyledEditForm = styled(EditForm)`
 `;
 
 const EditPage = () => {
+  const router = useRouter();
+  const dispatch = useDispatch();
   const diary = useSelector((state: RootState) => state.diary);
 
   return (
     <StyledLayout>
       <Heading.Text1 text="てつどうを記録する" as="h2" />
-      <StyledEditForm diary={diary} />
+      <StyledEditForm
+        diary={diary}
+        onSubmit={(title, body) => {
+          if (!diary) {
+            dispatch(createDraft({ id: undefined, title, body }));
+          } else {
+            dispatch(createDraft({ id: diary.id, title, body }));
+          }
+          if (body.length > 0) {
+            router.push("/preview");
+          }
+        }}
+      />
     </StyledLayout>
   );
 };
